+++ package.json
@@ -14,6 +14,7 @@
     "axios": "^1.6.7",
     "classnames": "^2.5.1",
     "next": "^12.3.4",
+    "next-i18next": "^15.2.0",
     "react": "^18",
     "react-calendar": "^4.8.0",
     "react-datepicker": "^6.4.0",
