+++ src/components/molecules/landing/landingCarousel/index.tsx
@@ -1,6 +1,5 @@
 import { TouchEventHandler, useEffect, useRef, useState } from 'react';
 import styles from './LandingCarousel.module.scss';
-import { StaticImageData } from 'next/image';
 import Image from 'next/image';
 import ForwardArrowIcon from '@/../../public/assets/icon/forwardArrowIcon.svg';
 import LeftArrowIcon from '@/../../public/assets/icon/leftArrowIcon.svg';
@@ -47,8 +46,6 @@ const LandingCarousel = ({ carouselList }: LandingCarouselProps) => {
   }, [carouselList, currIndex, carouselElement]);
 
   useEffect(() => {
-    window.addEventListener('onLoad', handleResize);
-    document.addEventListener('resize', handleResize);
     window.addEventListener('resize', handleResize);
   }, []);
